test(do): split relation-predicates bench into isolated per-shape files
The single file seeded two module-level writers in beforeAll; CodSpeed's
instrumented runner re-runs the suite, so the seed ran twice over the same
engines ("unique constraint violation on users"). Split into one file per
relation shape (to-one is, to-many some, multi-hop, no-relation), and build +
seed each file's writers FRESH inside its own beforeAll (via
`relation-predicates.shared.ts`'s `makeSeededRelationWriter`) so a repeated run
always starts from a clean engine. The push-down/semijoin comparison stays
intact per file.
