+++ packages/do/__bench__/relation-predicates-no-relation.bench.ts
@@ -1,0 +1,22 @@
+import { beforeAll, bench, describe } from "vitest";
+
+import type { DatabaseWriterLike } from "../src/ctx-db";
+import { makeSeededRelationWriter } from "./relation-predicates.shared";
+
+/**
+ * No-relation fast path — a flat predicate on a relation-bearing table must NOT
+ * pay for the feature beyond one synchronous `containsRelationPredicate` scan of
+ * `where`. Pins the overhead of the relation-predicate machinery on the common
+ * read that never crosses a relation.
+ */
+let writer: DatabaseWriterLike;
+
+beforeAll(async () => {
+    writer = await makeSeededRelationWriter("always");
+});
+
+describe("no-relation fast path — containsRelationPredicate overhead", () => {
+    bench("flat where on a relation-bearing table (no relation predicate)", async () => {
+        await writer.findMany("messages", { limit: 50, where: { body: { contains: "99" } } });
+    });
+});
